+++ src/components/todo/todoCreate/todoCreate.jsx
@@ -4,29 +4,19 @@ import styled from 'styled-components';
 import InputText from '../../input/inputText';
 import Button from '../../button/button';
 
-import { useTodoDispatch, useTodoNextId } from '../../../context/todoContext';
-
-function TodoCreate() {
+function TodoCreate({ createTodo }) {
     const [value, setValue] = useState('');
-
-    const dispatch = useTodoDispatch();
-    const nextId = useTodoNextId();
 
     const onChange = (e) => setValue(e.target.value);
     const onSubmit = (e) => {
         e.preventDefault();
 
-        dispatch({
-            type: 'CREATE',
-            todo: {
-                id: nextId.current,
-                text: value,
-                done: false,
-            },
-        });
+        const todoData = {
+            todo: value,
+        };
 
+        createTodo(todoData);
         setValue('');
-        nextId.current += 1;
     };
 
     return (
@@ -36,8 +26,9 @@ function TodoCreate() {
                 placeholder="할 일을 입력 후, Enter 를 누르세요"
                 onChange={onChange}
                 value={value}
+                data-testid="new-todo-input"
             />
-            <Button type="button" onClick={onSubmit}>
+            <Button onClick={onSubmit} data-testid="new-todo-add-button">
                 추가
             </Button>
         </TodoForm>
